+++ flutter_plugin_qpos/CHANGELOG.md
@@ -3,3 +3,4 @@
 * 1. Add callbacks for bluetooth power on and power off on flutter.
 * 2. fix scan bluetooth twice issue(IOS).
 * 3. Add function for mifare operation: pollOnMifareCard; authenticateMifareCard; operateMifareCardData; readMifareCard; writeMifareCard; finishMifareCard
+* 4. Add function setBuzzerStatus, doSetBuzzerOperation